The saturation slider (the right side/fully saturated side) should reflect the color wheel at it's defaults, not it's current settings.
Pratically, this means that setting saturation and brightness to their lowest settings and clicking on the hue wheel should still show the color that _would be there_ at the default (sensible?) settings.

This way the user can see the actual color at that location on the wheel even if the brightness is way down, etc. This is more like what photoshop does, which I think is a good thing.